+++ webpack.config.js
@@ -1,9 +1,14 @@
 /* global __dirname */
-var path = require('path');
-var dir_js = path.resolve(__dirname, 'react_components');
-var dir_build = path.resolve(__dirname, 'build');
 
-module.exports = {
+const path = require('path');
+
+const dir_js = path.resolve(__dirname, 'react_components');
+const dir_build = path.resolve(__dirname, 'build');
+const dir_dist = path.resolve(__dirname, 'dist');
+const dir_node_modules = path.resolve(__dirname, 'node_modules');
+
+const config = {
+  target: 'node',
   entry: path.resolve(dir_js, 'index.js'),
   output: {
     path: dir_build,
@@ -19,10 +24,12 @@ module.exports = {
       {
         use: 'react-hot-loader/webpack',
         test: dir_js,
+        exclude: dir_node_modules,
       },
       {
         use: 'babel-loader',
         test: dir_js,
+        exclude: dir_node_modules,
       },
     ],
   },
@@ -42,3 +49,10 @@ module.exports = {
   devtool: 'source-map',
   mode: 'development',
 };
+
+module.exports = (env, argv) => {
+  if (argv.mode === 'production') {
+    config.output.path = dir_dist;
+  }
+  return config;
+}